+++ ErsatzTV.Infrastructure/Search/LuceneSearchIndex.cs
@@ -34,6 +34,7 @@ public sealed class LuceneSearchIndex : ISearchIndex
     internal const string SortTitleField = "sort_title";
     internal const string GenreField = "genre";
     internal const string TagField = "tag";
+    internal const string TagFullField = "tag_full";
     internal const string PlotField = "plot";
     internal const string LibraryNameField = "library_name";
     internal const string LibraryIdField = "library_id";
@@ -110,7 +111,7 @@ public sealed class LuceneSearchIndex : ISearchIndex
         return Task.FromResult(directoryExists && fileExists);
     }
 
-    public int Version => 44;
+    public int Version => 45;
 
     public async Task<bool> Initialize(
         ILocalFileSystem localFileSystem,
@@ -462,6 +463,7 @@ public sealed class LuceneSearchIndex : ISearchIndex
                 foreach (Tag tag in metadata.Tags)
                 {
                     doc.Add(new TextField(TagField, tag.Name, Field.Store.NO));
+                    doc.Add(new StringField(TagFullField, tag.Name, Field.Store.NO));
                 }
 
                 foreach (Studio studio in metadata.Studios)
@@ -649,6 +651,7 @@ public sealed class LuceneSearchIndex : ISearchIndex
                 foreach (Tag tag in metadata.Tags)
                 {
                     doc.Add(new TextField(TagField, tag.Name, Field.Store.NO));
+                    doc.Add(new StringField(TagFullField, tag.Name, Field.Store.NO));
                 }
 
                 foreach (Studio studio in metadata.Studios)
@@ -784,6 +787,7 @@ public sealed class LuceneSearchIndex : ISearchIndex
                 foreach (Tag tag in metadata.Tags)
                 {
                     doc.Add(new TextField(TagField, tag.Name, Field.Store.NO));
+                    doc.Add(new StringField(TagFullField, tag.Name, Field.Store.NO));
                 }
 
                 AddMetadataGuids(metadata, doc);
@@ -928,6 +932,7 @@ public sealed class LuceneSearchIndex : ISearchIndex
                 foreach (Tag tag in metadata.Tags)
                 {
                     doc.Add(new TextField(TagField, tag.Name, Field.Store.NO));
+                    doc.Add(new StringField(TagFullField, tag.Name, Field.Store.NO));
                 }
 
                 foreach (Studio studio in metadata.Studios)
@@ -1076,6 +1081,7 @@ public sealed class LuceneSearchIndex : ISearchIndex
                 foreach (Tag tag in metadata.Tags)
                 {
                     doc.Add(new TextField(TagField, tag.Name, Field.Store.NO));
+                    doc.Add(new StringField(TagFullField, tag.Name, Field.Store.NO));
                 }
 
                 foreach (Studio studio in metadata.Studios)
@@ -1186,6 +1192,7 @@ public sealed class LuceneSearchIndex : ISearchIndex
                 foreach (Tag tag in metadata.Tags)
                 {
                     doc.Add(new TextField(TagField, tag.Name, Field.Store.NO));
+                    doc.Add(new StringField(TagFullField, tag.Name, Field.Store.NO));
                 }
 
                 foreach (Studio studio in metadata.Studios)
@@ -1274,6 +1281,7 @@ public sealed class LuceneSearchIndex : ISearchIndex
                 foreach (Tag tag in metadata.Tags)
                 {
                     doc.Add(new TextField(TagField, tag.Name, Field.Store.NO));
+                    doc.Add(new StringField(TagFullField, tag.Name, Field.Store.NO));
                 }
 
                 foreach (Genre genre in metadata.Genres)
@@ -1345,6 +1353,7 @@ public sealed class LuceneSearchIndex : ISearchIndex
                 foreach (Tag tag in metadata.Tags)
                 {
                     doc.Add(new TextField(TagField, tag.Name, Field.Store.NO));
+                    doc.Add(new StringField(TagFullField, tag.Name, Field.Store.NO));
                 }
 
                 foreach (Genre genre in metadata.Genres)